app/requirements.txt
fastapi
uvicorn
httpx
pydantic-settings
google-genai